done with 4.20 - 4.23 and part4

+++ part4/blog/controllers/blogs.js
@@ -1,15 +1,12 @@
 import express from 'express';
 import Blog from '../models/blog.js';
-import User from '../models/user.js';
-import jwt from 'jsonwebtoken';
-import { SECRET } from '../utils/config.js';
 
 const blogRouter = express.Router();
 
 blogRouter.get('/', async (request, response) => {
     const blogs = await Blog
       .find({})
-      .populate('users', { username: 1, name: 1, id: 1 });
+      .populate('user', { username: 1, name: 1, id: 1 });
       
     response.json(blogs);
       
@@ -18,12 +15,7 @@ blogRouter.get('/', async (request, response) => {
 blogRouter.post('/', async (request, response) => {
     const {author, title, url, likes} = request.body;
 
-    const decodedToken = jwt.verify(request.token, SECRET);
-
-    if (!token || !decodedToken.id) {
-      return response.status(401).json({ error: 'token missing or invalid' })
-    }
-    const user = await User.findById(decodedToken.id);
+    const user = request.user;
 
     if (!title && !url) {
       return response.status(400).end();
@@ -47,7 +39,21 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+  const user = request.user;
+
+  const blog = await Blog.findById(request.params.id);
+
+  if (!blog) return response.status(404).json({error: 'bad blog id'});
+
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'unauthorized user' });
+  }
+
   await Blog.findByIdAndRemove(request.params.id);
+
+  user.blogs = user.blogs.filter(userblog => userblog.toString() !== blog._id.toString());
+  await user.save();
+
   response.status(204).end();
 })
